+++ plugins/voxelman/blockentity/utils.d
@@ -14,6 +14,23 @@ import voxelman.blockentity.blockentityaccess;
 import voxelman.world.storage.coordinates;
 import voxelman.world.storage.volume;
 
+enum BLOCK_ENTITY_FLAG = 1 << 15;
+enum BLOCK_INDEX_MASK = (1 << 15) - 1;
+
+bool isBlockEntity(BlockId blockId)
+{
+	enum n = BlockId.sizeof*8 - 1;
+	return blockId >> n; // highest bit
+}
+
+ushort blockIndexFromBlockId(BlockId blockId) {
+	return blockId & BLOCK_INDEX_MASK;
+}
+
+BlockId blockIdFromBlockIndex(ushort blockIndex) {
+	return blockIndex | BLOCK_ENTITY_FLAG;
+}
+
 
 alias BlockEntityMeshhandler = void function(
 	ref Appender!(ubyte[]) output,
@@ -23,6 +40,7 @@ alias BlockEntityMeshhandler = void function(
 
 alias SolidityHandler = Solidity function(Side side);
 alias EntityBoxHandler = Volume function(BlockWorldPos bwp, BlockEntityData data);
+alias EntityDebugHandler = void function(BlockWorldPos bwp, BlockEntityData data);
 Volume nullBoxHandler(BlockWorldPos bwp, BlockEntityData data)
 {
 	return Volume(bwp.xyz, ivec3(1,1,1), cast(ushort)bwp.w);
@@ -41,6 +59,7 @@ struct BlockEntityInfo
 	BlockEntityMeshhandler meshHandler;
 	SolidityHandler sideSolidity;
 	EntityBoxHandler boxHandler;
+	EntityDebugHandler debugHandler;
 	ubyte[3] color;
 	//bool isVisible = true;
 	size_t id;